#104: rename *IntegrationTest to *IT
