+++ src/test/integration/java/com/commercetools/config/CtpConfigStartupValidatorTestIT.java
@@ -55,7 +55,7 @@ import static org.mockito.Mockito.verify;
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class CtpConfigStartupValidatorTestIntegrationTest {
+public class CtpConfigStartupValidatorTestIT {
 
     @Autowired
     private CtpConfigStartupValidator ctpConfigStartupValidator;
