+++ src/test/integration/java/com/commercetools/payment/handler/CommercetoolsCreatePaymentWithShippingPreferenceIT.java
@@ -1,7 +1,7 @@
 package com.commercetools.payment.handler;
 
 import com.commercetools.Application;
-import com.commercetools.payment.PaymentIntegrationTest;
+import com.commercetools.payment.BasePaymentIT;
 import com.commercetools.pspadapter.facade.CtpFacade;
 import com.commercetools.pspadapter.facade.CtpFacadeFactory;
 import com.commercetools.pspadapter.facade.SphereClientFactory;
@@ -55,7 +55,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @AutoConfigureMockMvc
 // 1) completely wipe-out CTP project Payment, Cart, Order endpoints before the test cases
 @Import(OrdersCartsPaymentsCleanupConfiguration.class)
-public class CommercetoolsCreatePaymentWithShippingPreferenceIT extends PaymentIntegrationTest {
+public class CommercetoolsCreatePaymentWithShippingPreferenceIT extends BasePaymentIT {
 
     @Autowired
     private MockMvcAsync mockMvcAsync;
