+++ src/test/integration/java/com/commercetools/pspadapter/tenant/TenantConfigFactoryIT.java
@@ -14,7 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class TenantConfigFactoryIntegrationTest {
+public class TenantConfigFactoryIT {
 
     @Autowired
     private TenantConfigFactory tenantConfigFactory;
